Создание, обновление и загрузка LastUserSymptom

+++ app/src/main/java/org/blagodarie/ui/symptoms/DisplaySymptom.java
@@ -40,7 +40,7 @@ public final class DisplaySymptom
     @NonNull
     private ObservableBoolean mHaveNotSynced = new ObservableBoolean(false);
 
-    private volatile int mUserSymptomCount = 0;
+    private volatile long mUserSymptomCount = 0;
 
     @NonNull
     private ObservableBoolean mHighlight = new ObservableBoolean(false);
@@ -85,11 +85,11 @@ public final class DisplaySymptom
         return mHaveNotSynced;
     }
 
-    int getUserSymptomCount () {
+    long getUserSymptomCount () {
         return mUserSymptomCount;
     }
 
-    void setUserSymptomCount (final int userSymptomCount) {
+    void setUserSymptomCount (final long userSymptomCount) {
         this.mUserSymptomCount = userSymptomCount;
     }
